Develop

+++ internal/file/file.go
@@ -5,37 +5,66 @@ import (
 	"io"
 	"io/ioutil"
 	"net/http"
+	"net/url"
 	"os"
 	"path"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
+
+	"github.com/getsumio/getsum/internal/status"
 )
 
+//file interface
 type IFile interface {
 	Path() string
 	Data() ([]byte, error)
 	Fetch(timeout int) error
 	IsRemote() bool
+	Delete()
+	Reset()
 }
 
+//file struct
 type File struct {
 	path        string
 	data        []byte
 	Url         string
-	Status      string
-	StatusValue string
+	Status      *status.Status
 	Size        int64
+	Proxy       string
+	StoragePath string
 }
 
+//file location on local host
+//if file is remote file something like http/https
+//path will be present after file is downloaded
+//so you need to call Fetch method first
 func (f *File) Path() string {
 	return f.path
 }
 
-func (f *File) IsRemote() bool {
-	return strings.HasPrefix(f.Url, "http") || strings.HasPrefix(f.Url, "ftp")
+//simply removes file if its already fetched
+func (f *File) Delete() {
+	if f.path != "" {
+		os.Remove(f.path)
+	}
 }
 
+//checks file is not on local
+//TODO add file:/// support
+func (f *File) IsRemote() bool {
+	return strings.HasPrefix(f.Url, "http")
+}
+
+//a bit of ugly but resets global variable
+//check variable comment for details
+func (f *File) Reset() {
+	fetchedSize = -1
+}
+
+//read the file data in bytes
 func (f *File) Data() ([]byte, error) {
 	if f.path == "" {
 		return nil, errors.New("File not fetched yet you need to first call Fetch()")
@@ -44,11 +73,14 @@ func (f *File) Data() ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	f.Status = "ALLOCATED"
+	f.Status.Type = status.ALLOCATED
 	return bytes, nil
 
 }
 
+//validates file
+//if remote fetches file
+//sets path location forthe stored file
 func (f *File) Fetch(timeout int) error {
 	err := validateUrl(f)
 	if err != nil {
@@ -60,26 +92,51 @@ func (f *File) Fetch(timeout int) error {
 	} else {
 		return fetchRemote(f, timeout)
 	}
-	f.Status = "FETCHED"
 
 	return nil
 }
 
+//TODO: sync.ONCE maybe better?
+var mux sync.Mutex
+
+//in case of user runs multiple checksum calculations
+//on same machine there is no point to
+//download file per routine
+//so first routine takes the leads and downloads file
+//then other routines check this variable
+//if present they use existing path
+var fetchedSize int64 = -1
+
+//fetches file remotely
+//unless not timedout sets details and path
 func fetchRemote(f *File, timeout int) error {
 
+	mux.Lock()
+	defer mux.Unlock()
+	filename := path.Base(f.Url)
+	if f.StoragePath != "" {
+		filename = strings.Join([]string{f.StoragePath, filename}, "/")
+	}
+	if fetchedSize > 0 { //another process already fetched
+		f.Size = fetchedSize
+		f.path = filename
+		f.Status.Type = status.FETCHED
+		return nil
+	}
 	err := validateRemote(f)
 	if err != nil {
 		return err
 	}
 
-	filename := path.Base(f.Url)
 	out, err := os.Create(filename)
 	if err != nil {
 		return err
 	}
 	defer out.Close()
 
-	header, err := http.Head(f.Url)
+	client := getHttpClient(f, timeout)
+
+	header, err := client.Head(f.Url)
 	if err != nil {
 		return err
 	}
@@ -92,14 +149,12 @@ func fetchRemote(f *File, timeout int) error {
 	f.path = filename
 
 	quit := make(chan bool)
+	defer close(quit)
 	go downloadFile(quit, f)
-
-	client := http.Client{
-		Timeout: time.Duration(timeout) * time.Second,
-	}
 
 	resp, err := client.Get(f.Url)
 	if err != nil {
+		quit <- true
 		return err
 	}
 	defer resp.Body.Close()
@@ -111,20 +166,27 @@ func fetchRemote(f *File, timeout int) error {
 	_, err = io.Copy(out, resp.Body)
 	quit <- true
 
+	f.Status.Type = status.FETCHED
+	fetchedSize = f.Size
 	return err
 }
 
+//only validates file since its already hosted
+//see validation for details
 func fetchLocal(f *File) error {
 
 	err := validateLocal(f)
 	if err != nil {
+		f.Status.Type = status.ERROR
 		return err
 	}
 	f.path = f.Url
+	f.Status.Type = status.FETCHED
 	return nil
 
 }
 
+//removes file
 func remove(fileOrDir string) error {
 	err := os.Remove(fileOrDir)
 	if err != nil {
@@ -132,4 +194,23 @@ func remove(fileOrDir string) error {
 	}
 	return nil
 
+}
+
+//return http client wrapped by
+//proxy settings
+//as well as timeout value
+func getHttpClient(f *File, timeout int) *http.Client {
+	proxyUrl := http.ProxyFromEnvironment
+	if f.Proxy != "" {
+		proxy, _ := url.Parse(f.Proxy)
+		proxyUrl = http.ProxyURL(proxy)
+	}
+	tr := &http.Transport{
+		Proxy: proxyUrl,
+	}
+	client := &http.Client{
+		Transport: tr,
+		Timeout:   time.Duration(timeout) * time.Second,
+	}
+	return client
 }
